requirements.txt
torch
fastapi
uvicorn
tqdm
transformers
onnxruntime-gpu
sentencepiece
